+++ rtf_vis_tool/src/result_metrics/multiview_optimizer_metrics.json
@@ -1,0 +1,113 @@
+{
+    "rotation_averaging_angle_deg": {
+        "median_error": 2.915834376502234,
+        "min_error": 3.5338831558100146e-15,
+        "max_error": 4.913298819627427,
+        "errors_list": [
+            3.5338831558100146e-15,
+            0.4189390807955098,
+            1.3680711437638018,
+            1.740142833326435,
+            2.326309452558151,
+            2.7365251162847435,
+            3.0951436367197247,
+            3.4976886271627943,
+            3.987135840751106,
+            4.319320628556882,
+            4.527332144249488,
+            4.913298819627427
+        ]
+    },
+    "translation_averaging_distance": {
+        "median_error": 0.869501069185224,
+        "min_error": 1.1662709904749775e-16,
+        "max_error": 1.787777520888663,
+        "errors_list": [
+            1.1662709904749775e-16,
+            0.12977697982394862,
+            0.18791495404222752,
+            0.3399420179543568,
+            0.5627555386616466,
+            0.8153359207884234,
+            0.9236662175820245,
+            1.0155584553575243,
+            1.3563636193333823,
+            1.3991455213226334,
+            1.496861225536501,
+            1.787777520888663
+        ]
+    },
+    "translation_to_direction_angle_deg": {
+        "median_error": 1.0143640194468362,
+        "min_error": 0.058668522493239145,
+        "max_error": 3.4938676701267446,
+        "errors_list": [
+            0.4227410060090331,
+            0.854989442406458,
+            1.0178511677084214,
+            1.71762661967673,
+            0.7049662203545051,
+            1.5674019335630685,
+            0.5119218104211672,
+            0.6374758751297421,
+            0.7991944941268373,
+            1.6083236181839677,
+            1.6568487783824477,
+            0.585566626206865,
+            1.1540807943682194,
+            1.7291633486060312,
+            2.332190768271201,
+            1.8315879620028972,
+            1.2367845985802495,
+            0.835769996417073,
+            0.9788627471956846,
+            1.412666524073752,
+            1.1146066656050282,
+            0.18004995311515387,
+            1.9052495233506244,
+            3.2732213993452395,
+            0.5244910867346038,
+            1.850816831817249,
+            0.5921218454525566,
+            1.2392327025292718,
+            1.4012407090589316,
+            1.4277138428754068,
+            0.9085557406242682,
+            1.137452077743756,
+            2.295812310239778,
+            0.980739501307924,
+            0.45264158293994644,
+            0.5572179546758942,
+            1.211215242104992,
+            1.3156149961198047,
+            0.4170478056370364,
+            1.6209626684559593,
+            1.6711402403518472,
+            1.0417064163389012,
+            0.9061353673823842,
+            0.058668522493239145,
+            1.0108768711852512,
+            0.2886328401245232,
+            3.4938676701267446,
+            0.34119851946695834,
+            1.4726601788433051,
+            0.21846189416572664,
+            0.7828002563493252,
+            0.6323237333470062,
+            2.366197832170407,
+            0.8376104496824853,
+            0.7237843805877027,
+            1.5336967795926593,
+            0.5943483218938266,
+            1.6649055411671383,
+            1.1249348235848071,
+            0.3820715777537737,
+            0.5875635849009108,
+            0.08896187569340668,
+            2.039001518155087,
+            0.4315152233837527,
+            1.4817637364506395,
+            0.9496869446321287
+        ]
+    }
+}
